feat: Add Pokemon card component and update usePokeApi.js & PokemonCheckList.js
Create a Pokemon card component that can display the results returned by the usePokeApi hook. The hook now returns {sprites, types, game_indices, id} for each Pokemon.

Pokemon checklist now displays the user's selected Pokemon. The checklist is updated whenever the user adds a new Pokemon from the search results or removes an existing item. The search results are fetched from the PokeAPI using an input in the Pokemon checklist component.

Also added some CSS for the Pokemon checklist to differentiate when PokemonCard is a search result or added to the check list.

Note: Design and CSS is still a work in progress